+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Methods to the shared class that enable easy logging and profiling
+- Inheritance affordances for creating Async components as per [Dimitre's GrasshopperAsyncComponent](https://github.com/specklesystems/GrasshopperAsyncComponent/)
 
 
 ## [0.10.0] - 2021-04-11